Move to stats package
